requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
python-decouple==3.8
mysqlclient==2.0.3
mysql-connector-python==8.3.0
